requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
flask-restx==1.3.0
flask-limiter==3.5.0
python-dotenv==1.0.1
pytest==8.2.0